test jest

+++ test/setupTestDB.ts
@@ -1,19 +1,27 @@
 import mongoose from 'mongoose';
 import connectDB from '../backend/config/connectDB';
 
-const MONGO_URI = String(process.env.MONGO_URI_TEST!);
-
 export const connectTestDB = async () => {
     try {
-        await connectDB(MONGO_URI);
+        if (mongoose.connection.readyState === 0) {
+            await connectDB(process.env.MONGO_URI_TEST!);
+            console.log('Connected to TEST database');
+        }
     } catch (error) {
-        console.error('Database connection error:', error);
-        throw error;
+        console.error('Test DB connection error:', error);
+        process.exit(1);
     }
 };
 
 export const disconnectTestDB = async () => {
-    await mongoose.connection.dropDatabase();
-    await mongoose.disconnect();
-    console.log('MongoDB Disconnected');
+    try {
+        if (mongoose.connection.readyState !== 0) {
+            await mongoose.connection.dropDatabase();
+            await mongoose.disconnect();
+            console.log('Disconnected from TEST database');
+        }
+    } catch (error) {
+        console.error('Test DB disconnection error:', error);
+        process.exit(1);
+    }
 };
